РК-3

+++ internal/delivery/cart/dto.go
@@ -65,9 +65,9 @@ func productResponseFromModel(cartProduct model.CartProduct) GetProductResponse 
 }
 
 type orderData struct {
-	TotalItems     uint16             `json:"total_items"`
+	TotalItems     uint32             `json:"total_items"`
 	TotalWeight    float32            `json:"total_weight"`
-	FinalPrice     float32            `json:"final_price"`
+	FinalPrice     uint32             `json:"final_price"`
 	Currency       string             `json:"currency"`
 	PaymentMethods []paymentMethod    `json:"payment_methods"`
 	Recipient      recipientInfo      `json:"recipient"`
@@ -93,8 +93,8 @@ type deliveryDateInfo struct {
 
 type productResponse struct {
 	Title    string  `json:"product_name"`
-	Price    float32 `json:"product_price"`
-	Quantity uint16  `json:"quantity"`
+	Price    uint32  `json:"product_price"`
+	Quantity uint32  `json:"quantity"`
 	Image    string  `json:"product_image"`
 	Weight   float32 `json:"weight"`
 	URL      string  `json:"url"`
